rename geo-compose -> geoview-compose (#361)
* rename geo-compose -> geoview-compose

* update doc

* update reference in micro app READMEs

* update logger warning,
update package name

* update namespace in build.gradle
