+++ microapps/FloorFilterApp/app/build.gradle.kts
@@ -63,7 +63,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile::class).all {
 
 dependencies {
     implementation(project(":indoors"))
-    implementation(project(":geo-compose"))
+    implementation(project(":geoview-compose"))
     implementation(arcgis.mapsSdk)
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.bundles.composeCore)
