+++ microapps/MapViewGeometryEditorApp/README.md
@@ -9,4 +9,4 @@ This micro-app demonstrates the use of `GeometryEditor` and `GraphicsOverlay` wi
 The application starts with a `GeometryEditor` set up and not enabled. Use the switch in the app bar to start/stop the geometry editor session.
 Use the overflow action button in the app bar to choose between different options to undo, redo, or clear all `Graphics` from the `GraphicsOverlay`.
 
-For more information on the composable `MapView` component and how it works, see it's [Readme](../../toolkit/geo-compose/README.md).
+For more information on the composable `MapView` component and how it works, see it's [Readme](../../toolkit/geoview-compose/README.md).
