+++ microapps/MapViewIdentifyApp/app/src/main/java/com/arcgismaps/toolkit/mapviewidentifyapp/screens/MainScreen.kt
@@ -50,7 +50,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.arcgismaps.toolkit.geocompose.MapView
+import com.arcgismaps.toolkit.geoviewcompose.MapView
 import com.arcgismaps.toolkit.mapviewidentifyapp.R
 
 private val detailsHeight = 200.dp
