+++ microapps/MapViewInsetsApp/README.md
@@ -9,4 +9,4 @@ This micro-app demonstrates the use of `Insets` with a composable `MapView`.
 The application starts with a `Map` with no insets specifed. Use the corresponding text fields to specify left, right, top or bottom insets.
 The attribution bar at the bottom reacts to the inset values specified. The `Reset Insets` button resets all the insets.
 
-For more information on the composable `MapView` component and how it works, see it's [Readme](../../toolkit/geo-compose/README.md).
+For more information on the composable `MapView` component and how it works, see it's [Readme](../../toolkit/geoview-compose/README.md).
