+++ microapps/MapViewInsetsApp/app/build.gradle.kts
@@ -80,7 +80,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile::class).all {
 }
 
 dependencies {
-    implementation(project(":geo-compose"))
+    implementation(project(":geoview-compose"))
     implementation(arcgis.mapsSdk)
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.bundles.composeCore)
