+++ microapps/SceneViewAnalysisOverlayApp/README.md
@@ -8,4 +8,4 @@ This micro-app demonstrates the use of composable `SceneView's` `analysisOverlay
 
 The application starts with a SceneView and displays a scene of Brest, France. Use the switch to turn on and off a viewshed analysis from the viewpoint of the cyan diamond marker.
 
-For more information on the composable `SceneView` component and how it works, see it's [Readme](../../toolkit/geo-compose/README.md).
+For more information on the composable `SceneView` component and how it works, see it's [Readme](../../toolkit/geoview-compose/README.md).
