+++ microapps/SceneViewSetViewpointApp/README.md
@@ -10,5 +10,5 @@ This micro-app demonstrates setting the viewpoint on a composable `SceneView`.
 
 The application starts with a SceneView and displays a scene. Use the overflow action button in the app bar to choose between different methods of setting a viewpoint on the composable SceneView. Each operation will set the viewpoint to a different location.
 
-For more information on the composable `SceneView` component and how it works, see it's [Readme](../../toolkit/geo-compose/README.md).
+For more information on the composable `SceneView` component and how it works, see it's [Readme](../../toolkit/geoview-compose/README.md).
