+++ settings.gradle.kts
@@ -77,7 +77,7 @@ include(":kdoc")
 include(":composable-map")
 include(":indoors")
 include(":floor-filter-app")
-include(":geo-compose")
+include(":geoview-compose")
 include(":map-view-location-display-app")
 include(":map-view-insets-app")
 include(":map-view-geometry-editor-app")
@@ -97,7 +97,7 @@ project(":bom").projectDir = File(rootDir, "bom")
 project(":composable-map").projectDir = File(rootDir, "toolkit/composable-map")
 project(":indoors").projectDir = File(rootDir, "toolkit/indoors")
 project(":floor-filter-app").projectDir = File(rootDir, "microapps/FloorFilterApp/app")
-project(":geo-compose").projectDir = File(rootDir, "toolkit/geo-compose")
+project(":geoview-compose").projectDir = File(rootDir, "toolkit/geoview-compose")
 project(":map-view-location-display-app").projectDir = File(rootDir, "microapps/mapviewlocationdisplayapp/app")
 project(":map-view-insets-app").projectDir = File(rootDir, "microapps/mapviewinsetsapp/app")
 project(":map-view-geometry-editor-app").projectDir = File(rootDir, "microapps/mapviewgeometryeditorapp/app")
