+++ toolkit/geoview-compose/src/main/java/com/arcgismaps/toolkit/geoviewcompose/SceneView.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.arcgismaps.toolkit.geocompose
+package com.arcgismaps.toolkit.geoviewcompose
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
@@ -112,7 +112,7 @@ import java.time.Instant
  * @param onTwoPointerTap lambda invoked when a user taps two pointers on the composable SceneView
  * @param onPan lambda invoked when a user drags a pointer or pointers across composable SceneView
  * @param onDrawStatusChanged lambda invoked when the draw status of the composable SceneView is changed
- * @sample com.arcgismaps.toolkit.geocompose.samples.SceneViewSample
+ * @sample com.arcgismaps.toolkit.geoviewcompose.samples.SceneViewSample
  * @see
  * - <a href="https://developers.arcgis.com/kotlin/scenes-3d/tutorials/display-a-scene/">Display a scene tutorial</a>
  * - <a href="https://developers.arcgis.com/kotlin/scenes-3d/tutorials/display-a-web-scene/">Display a web scene tutorial</a>
@@ -456,7 +456,7 @@ private fun ViewpointHandler(
 /**
  * Contains default values for the SceneView.
  *
- * @see com.arcgismaps.toolkit.geocompose.SceneView
+ * @see com.arcgismaps.toolkit.geoviewcompose.SceneView
  * @since 200.4.0
  */
 public object SceneViewDefaults {
